fix: gradient placement
